Chose different algorithm from same source, width-shooting. Reworked code to work with this algorithm. (Check description))
The reason i chose for a different form of the algorithm is because the vertical application of the algorithm is very hard to comprehend and not explained in further detail in my source. Singleplayer will still hold its functionality with vertical placed ships. (Game versus AI will only place ships horizontally so the algorithm has a chance of winning.)