Final commit (Check description)
AI: commented battlefield print.

PROJECT CODE IS NOW COMPLETE

Finally:
Did not create pytest code because it is almost impossible to implement. Should I have chosen to implement it, I would have to collect all functions in a single testfile and run it all which is practically unneccesary and you get the same result by running the main code.